The nascar scraper has been extracted from the nascar-irc app to its own module. This is just an early draft of this, somewhat POC. A refactoring/cleanup is now desperately needed, and will be the next step.
nascar-stats-scraper however is on its way to being modular enough that it could be used in other situations. LeaderboardClient takes a URL and refresh_interval and an event is fired upon a successful update of the data set and all indexes.

We'll want to add some more events along the way, probably. And have them be named saner.

In any case, rough draft of extraction: Complete.

(And yes the damn thing still runs and works.)