nascar-stats-scraper updateInterval method added
The updateInterval method will allow the calling program to request a
change to the refresh interval (how fast we hit the NASCAR JSON).

It also performs some basic sanity checking on the passed value and
will emit a WARNING event, and set the interval to 60sec, if non-sane
values are passed.