Moving tests to tests/ and adding lap-by-lap
bin/ was not a good place for our .json test files, so they've been
moved to tests/ for now.

Also added some truncated lapbylap.json files that grow in their
numbered order. So you can host them on the web and rotate them at a
given URL to test the lap-by-lap eventer which is to come.