LapByLapClient added to nascar-stats-scraper
The new LapByLapClient works similar to LeaderboardClient. It leverages
the passed URL and refresh interval to poll the lap-by-lap JSON data.

An array is used to denote information about seen lap labels, so that we don't
re-post lap information to the channel. This data is also logged to the file,
and a read attempt is made at startup.

This client is EXTREMELY ROUGH at this point, needs to be cleaned up
significantly, code needs to be correctly styled, error handling added, etc..