Fix driver status delta message to make more sense
The driver status message would show "currently leading" while
waiting for the race to start. Fixed this to handle pre-race delta
of 0.

I am not sure if there are cases where delta might be 0 for a position
greater than 1 but have a placeholder output of "i don't know" for now.